+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Fixed**
 
 - Allow spaces in the parser function name in the MediaWiki mode
+- Allow special characters in the anchor of a template name in the MediaWiki mode
 
 ## 2.8.0